:construction:reset minikin to UTF16

+++ src/gui/text/minikin/include/minikin/Measurement.h
@@ -22,10 +22,10 @@
 
 namespace minikin {
 
-float getRunAdvance(const float* advances, const uint32_t* buf, size_t start, size_t count,
+float getRunAdvance(const float* advances, const uint16_t* buf, size_t start, size_t count,
                     size_t offset);
 
-size_t getOffsetForAdvance(const float* advances, const uint32_t* buf, size_t start, size_t count,
+size_t getOffsetForAdvance(const float* advances, const uint16_t* buf, size_t start, size_t count,
                            float advance);
 
 }  // namespace minikin
